fix crashing on big frames (#1)
GetMessageLength and GetMessageSequence uint8 -> uint

types are overly strict, and mess with arithmetic
e.g. uint8_var + literal -> uint8
this can result in overflows/wrap arounds
there's no reason to have uint8, so bump to uint

+++ frame.go
@@ -34,8 +34,8 @@ import (
 // Frame represents a MAVLink frame containing a MAVLink message
 type Frame interface {
 	GetVersion() int
-	GetMessageLength() uint8
-	GetMessageSequence() uint8
+	GetMessageLength() uint
+	GetMessageSequence() uint
 	GetSenderSystemID() uint8
 	GetSenderComponentID() uint8
 	GetMessageBytes() []byte
@@ -59,13 +59,13 @@ func (frame FrameV1) GetVersion() int {
 }
 
 // GetMessageLength returns the length of the message contained in the Frame
-func (frame FrameV1) GetMessageLength() uint8 {
-	return frame[1]
+func (frame FrameV1) GetMessageLength() uint {
+	return uint(frame[1])
 }
 
 // GetMessageSequence the sequence number of the message contained in the Frame
-func (frame FrameV1) GetMessageSequence() uint8 {
-	return frame[2]
+func (frame FrameV1) GetMessageSequence() uint {
+	return uint(frame[2])
 }
 
 // GetMessageBytes returns the message contained in the Frame as a byte array
@@ -90,7 +90,7 @@ func (frame FrameV1) GetMessageID() uint32 {
 
 // GetChecksum returns the checksum of the Frame contents
 func (frame FrameV1) GetChecksum() uint16 {
-	start := int(frame.GetMessageLength()) + 6
+	start := frame.GetMessageLength() + 6
 	end := start + 2
 	return binary.LittleEndian.Uint16(frame[start:end])
 }
@@ -98,7 +98,7 @@ func (frame FrameV1) GetChecksum() uint16 {
 // GetChecksumInput returns the contents of the Frame used to calculate the checksum
 func (frame FrameV1) GetChecksumInput() []byte {
 	start := 1
-	end := int(frame.GetMessageLength()) + 6
+	end := frame.GetMessageLength() + 6
 	return frame[start:end]
 }
 
@@ -134,8 +134,8 @@ func (frame FrameV2) GetVersion() int {
 }
 
 // GetMessageLength returns the length of the message contained in the Frame
-func (frame FrameV2) GetMessageLength() uint8 {
-	return frame[1]
+func (frame FrameV2) GetMessageLength() uint {
+	return uint(frame[1])
 }
 
 // GetIncompatibilityFlags returns the incompatibility flags for the Frame
@@ -149,8 +149,8 @@ func (frame FrameV2) GetCompatibilityFlags() uint8 {
 }
 
 // GetMessageSequence the sequence number of the message contained in the Frame
-func (frame FrameV2) GetMessageSequence() uint8 {
-	return frame[4]
+func (frame FrameV2) GetMessageSequence() uint {
+	return uint(frame[4])
 }
 
 // GetSenderSystemID returns the ID of the system that sent the Frame
@@ -174,14 +174,14 @@ func (frame FrameV2) GetMessageID() uint32 {
 
 // GetMessageBytes returns the message contained in the Frame as a byte array
 func (frame FrameV2) GetMessageBytes() []byte {
-	start := 10
-	end := start + int(frame.GetMessageLength())
+	start := uint(10)
+	end := start + frame.GetMessageLength()
 	return frame[start:end]
 }
 
 // GetChecksum returns the checksum of the Frame contents
 func (frame FrameV2) GetChecksum() uint16 {
-	start := int(frame.GetMessageLength()) + 10
+	start := frame.GetMessageLength() + 10
 	end := start + 2
 	return binary.LittleEndian.Uint16(frame[start:end])
 }
@@ -189,7 +189,7 @@ func (frame FrameV2) GetChecksum() uint16 {
 // GetChecksumInput returns the contents of the Frame used to calculate the checksum
 func (frame FrameV2) GetChecksumInput() []byte {
 	start := 1
-	end := int(frame.GetMessageLength()) + 10
+	end := frame.GetMessageLength() + 10
 	return frame[start:end]
 }
